// src/types/partials/ConsentProvision.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TPeriod } from '../partials/Period';
import { TConsentActor } from '../partials/ConsentActor';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TCoding } from '../partials/Coding';
import { TConsentData } from '../partials/ConsentData';

export type TConsentProvision = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    type?: String;
    period?: TPeriod;
    actor?: TConsentActor[];
    action?: TCodeableConcept[];
    securityLabel?: TCoding[];
    purpose?: TCoding[];
    class_?: TCoding[];
    code?: TCodeableConcept[];
    dataPeriod?: TPeriod;
    data?: TConsentData[];
    provision?: TConsentProvision[];
};